fix: move resources and learning content to home folder
